323. Number of Connected Components in an Undirected Graph
別のサイトで多分同じ問題を見つけたのでこちらで解きました：
https://neetcode.io/problems/count-connected-components

（問題文）

There is an undirected graph with n nodes. There is also an edges array, where edges[i] = [a, b] means that there is an edge between node a and node b in the graph.

The nodes are numbered from 0 to n - 1.

Return the total number of connected components in that graph.